# docs/api_specification.yaml
openapi: 3.0.3
info:
  title: PDF Smaller API
  description: |
    A comprehensive PDF compression service with user authentication, subscription management, 
    and bulk processing capabilities. This API provides endpoints for compressing PDF files, 
    managing user accounts, handling subscriptions, and processing multiple files in bulk.
    
    ## Features
    - Single and bulk PDF compression
    - User authentication with JWT tokens
    - Subscription management with Stripe integration
    - Usage tracking and rate limiting
    - File management and automatic cleanup
    - Comprehensive error handling and logging
    
    ## Authentication
    Most endpoints require authentication using JWT tokens. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your_jwt_token>
    ```
    
    ## Rate Limiting
    API endpoints are rate-limited based on user tier:
    - Anonymous users: 10 requests per minute
    - Free users: 20 requests per minute  
    - Premium users: 100 requests per minute
    - Pro users: 500 requests per minute
    
    ## Error Handling
    All errors follow a consistent format with error codes, messages, and request IDs for tracking.
    
  version: 1.0.0
  contact:
    name: PDF Smaller Support
    email: support@pdfsmaller.site
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.pdfsmaller.site
    description: Production server
  - url: https://staging-api.pdfsmaller.site
...
paths:
...
  /api/subscriptions:
    get:
      tags:
        - Subscriptions
      summary: Get user's subscription information
      description: Retrieve current subscription details and usage statistics
      operationId: getSubscriptionInfo
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Subscription information retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionInfo'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: No subscription found
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/subscriptions/plans:
    get:
      tags:
        - Subscriptions
      summary: Get available subscription plans
      description: Retrieve all available subscription plans and their features
      operationId: getAvailablePlans
      responses:
        '200':
          description: Plans retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlanList'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/subscriptions/create:
    post:
      tags:
        - Subscriptions
      summary: Create a new subscription
      description: Create a new subscription with Stripe payment processing
      operationId: createSubscription
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubscriptionCreate'
      responses:
        '201':
          description: Subscription created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionCreateResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: User already has a subscription
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/subscriptions/cancel:
    post:
      tags:
        - Subscriptions
      summary: Cancel user's subscription
      description: Cancel the user's active subscription
      operationId: cancelSubscription
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Subscription canceled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionCancelResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: No active subscription found
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/subscriptions/usage:
    get:
      tags:
        - Subscriptions
      summary: Get detailed usage statistics
      description: Retrieve detailed usage statistics and permissions for the user
      operationId: getUsageStatistics
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Usage statistics retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetailedUsageInfo'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/subscriptions/webhook:
    post:
      tags:
        - Subscriptions
      summary: Stripe webhook endpoint
      description: Handle Stripe webhook events for subscription updates
      operationId: stripeWebhook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Stripe webhook payload
      responses:
        '200':
          description: Webhook processed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
  schemas:
...
    UsageStats:
      type: object
      properties:
        compressions_today:
          type: integer
          description: Number of compressions performed today
          example: 5
        compressions_this_month:
          type: integer
          description: Number of compressions performed this month
          example: 45
        daily_limit:
          type: integer
          description: Daily compression limit
          example: 100
        monthly_limit:
          type: integer
          description: Monthly compression limit
          example: 1000
        files_processed_today:
          type: integer
          description: Number of files processed today
          example: 8
        total_size_compressed_mb:
          type: number
          format: float
          description: Total size compressed in MB
          example: 156.7
    Permissions:
      type: object
      properties:
        can_compress:
          type: boolean
          description: Can perform single file compression
          example: true
        can_bulk_compress:
          type: boolean
          description: Can perform bulk compression
          example: true
        can_use_api:
          type: boolean
          description: Can use API endpoints
          example: true
        max_file_size_mb:
          type: integer
          description: Maximum file size in MB
          example: 100
        max_bulk_files:
          type: integer
          description: Maximum files in bulk operation
          example: 20
        plan_name:
          type: string
          description: Current subscription plan
          example: "premium"

    # Bulk Compression Schemas
...
    Subscription:
      type: object
      properties:
        id:
          type: integer
          description: Subscription ID
          example: 789
        user_id:
          type: integer
          description: User ID
          example: 1
        plan_id:
          type: integer
          description: Plan ID
          example: 2
        plan_name:
          type: string
          description: Plan name
          example: "premium"
        stripe_subscription_id:
          type: string
          description: Stripe subscription ID
          example: "sub_1234567890"
        status:
          type: string
          enum: [active, canceled, past_due, unpaid]
          description: Subscription status
          example: "active"
        current_period_start:
          type: string
          format: date-time
          description: Current billing period start
          example: "2024-01-01T00:00:00Z"
        current_period_end:
          type: string
          format: date-time
          description: Current billing period end
          example: "2024-02-01T00:00:00Z"
        created_at:
          type: string
          format: date-time
          description: Subscription creation timestamp
          example: "2024-01-01T00:00:00Z"
    Plan:
      type: object
      properties:
        id:
          type: integer
          description: Plan ID
          example: 2
        name:
          type: string
          description: Plan name
          example: "premium"
        display_name:
          type: string
          description: Display name for the plan
          example: "Premium Plan"
        description:
          type: string
          description: Plan description
          example: "Advanced compression with bulk processing"
        price_monthly:
          type: number
          format: float
          description: Monthly price in USD
          example: 9.99
        price_yearly:
          type: number
          format: float
          description: Yearly price in USD
          example: 99.99
        features:
          type: object
          description: Plan features and limits
          properties:
            daily_compressions:
              type: integer
              example: 500
            bulk_processing:
              type: boolean
              example: true
            max_file_size_mb:
              type: integer
              example: 100
            max_bulk_files:
              type: integer
              example: 20
            api_access:
              type: boolean
              example: true
            priority_processing:
              type: boolean
              example: true
    SubscriptionInfo:
      type: object
      properties:
        subscription:
          $ref: '#/components/schemas/Subscription'
        usage:
          $ref: '#/components/schemas/UsageStats'
        status:
          type: object
          description: Subscription validation status
          properties:
            valid:
              type: boolean
              example: true
            permissions:
              $ref: '#/components/schemas/Permissions'
    PlanList:
      type: object
      properties:
        plans:
          type: array
          items:
            $ref: '#/components/schemas/Plan'
    SubscriptionCreate:
      type: object
      required:
        - plan_id
        - payment_method_id
      properties:
        plan_id:
          type: integer
          description: ID of the plan to subscribe to
          example: 2
        payment_method_id:
          type: string
          description: Stripe payment method ID
          example: "pm_1234567890"
        billing_cycle:
          type: string
          enum: [monthly, yearly]
          default: monthly
          description: Billing cycle preference
          example: "monthly"
    SubscriptionCreateResponse:
      type: object
      properties:
        message:
          type: string
          example: "Subscription created successfully"
        subscription:
          $ref: '#/components/schemas/Subscription'
        client_secret:
          type: string
          description: Stripe client secret for payment confirmation
          example: "pi_1234567890_secret_abc"
        subscription_id:
          type: string
          description: Stripe subscription ID
          example: "sub_1234567890"
    SubscriptionCancelResponse:
      type: object
      properties:
        message:
          type: string
          example: "Subscription canceled successfully"
        canceled_at:
          type: string
          format: date-time
          description: Cancellation timestamp
          example: "2024-01-15T10:30:00Z"
        current_period_end:
          type: string
          format: date-time
          description: End of current billing period
          example: "2024-02-01T00:00:00Z"
    DetailedUsageInfo:
      type: object
      properties:
        usage:
          $ref: '#/components/schemas/UsageStats'
        permissions:
          $ref: '#/components/schemas/Permissions'

    # Common Response Schemas
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Operation completed successfully"
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              description: Error code identifier
              example: "VALIDATION_ERROR"
            message:
              type: string
              description: Human-readable error message
              example: "Invalid input provided"
            details:
              type: object
              description: Additional error details
              example: {}
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
          example: "2024-01-15T10:30:00Z"
        request_id:
          type: string
          description: Unique request identifier for tracking
          example: "req_123456789"
  responses:
    BadRequest:
      description: Bad request - invalid input or missing required fields
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized - authentication required or invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...